Create Favorites.py
Favorites Database rough draft. Right now you can create the table, delete all favorites associated with a particular user, and get all of the entries for all users 100%. Currently working on the add_favorite because for some reason the thisdict1 dictionary works fine but the thisdict2 does not which is weird. Just wanted to go ahead and upload this so anybody can use it/see it if they like. Hopefully I will have the add_favorite issue fixed and the rest of the functionalities ready by next week.